feat: on-demand rendering on client-side
Tick events were removed, so for every frame we have to recalculate the current
state, based on the elapsed time, and the last known state.
This might cause performance issues if there is no communication happens
between the client and the server (we might have to simulate a *huge* amount of
time).